# sesi8-latihan/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sample API Spec for Orders",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Toni",
            "email": "toni@gmail.com"
        }
    },
    "basePath": "/",
    "paths": {
        "/people": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "person"
                ],
                "summary": "Get all people",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.GetAllPeopleSwagger"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "person"
                ],
                "summary": "Create people",
                "parameters": [
                    {
                        "description": "Create People",
                        "name": "People",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/params.Person"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.GetAllPeopleSwagger"
                        }
                    }
                }
            }
        },
        "/people/{id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "person"
                ],
                "summary": "Delete people",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "People ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.GetAllPeopleSwagger"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "person"
                ],
                "summary": "Update people",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "People ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update People",
                        "name": "People",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/params.Person"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.GetAllPeopleSwagger"
                        }
                    }
                }
            }
        },
        "/people_detail/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "person"
                ],
                "summary": "Get people",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "People ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.GetAllPeopleSwagger"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "params.Person": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "views.GetAllPeopleSwagger": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string",
                    "example": "GET_SUCCESS"
                },
                "payload": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/params.Person"
                    }
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        }
    }
}